[RENAME] 폴더명 booktest -> bookAPI로 변경 (API관련 코드만 따로 저장하기 위함)
