+++ src/main/java/com/mkl/book/bookAPI/DTO/BooksResponseDto.java
@@ -1,6 +1,6 @@
-package com.mkl.book.booktest.DTO;
+package com.mkl.book.bookAPI.DTO;
 
-import com.mkl.book.booktest.domain.Book;
+import com.mkl.book.bookAPI.domain.Book;
 import lombok.Data;
 
 import java.util.List;
