+++ src/main/java/com/mkl/book/bookAPI/controller/ReviewController.java
@@ -1,4 +1,4 @@
-package com.mkl.book.booktest.controller;
+package com.mkl.book.bookAPI.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
